requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
datasets==4.1.1
dill==0.4.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.9.0
huggingface-hub==0.35.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mysql-connector-python==9.4.0
networkx==3.4.2
numpy==2.1.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.3.2
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sympy==1.14.0
torch>=2.8.0
torchaudio>=2.8.0
torchvision>=0.19.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1